Cast string claims to int.
iat, exp, nbf should be parsed to json as an integer

+++ src/PayloadFactory.php
@@ -149,7 +149,7 @@ class PayloadFactory
      */
     public function iat()
     {
-        return Utils::now()->format('U');
+        return (int) Utils::now()->format('U');
     }
 
     /**
@@ -159,7 +159,7 @@ class PayloadFactory
      */
     public function exp()
     {
-        return Utils::now()->addMinutes($this->ttl)->format('U');
+        return (int) Utils::now()->addMinutes($this->ttl)->format('U');
     }
 
     /**
@@ -169,7 +169,7 @@ class PayloadFactory
      */
     public function nbf()
     {
-        return Utils::now()->format('U');
+        return (int) Utils::now()->format('U');
     }
 
     /**
